Ensure scanner is closed before finishing program

+++ src/main/java/library/constant/Constant.java
@@ -2,8 +2,8 @@ package library.constant;
 
 public final class Constant {
 
-    public static final long STANDARD_LOAN_PERIOD_WEEKS = 4;
-    public static final long AGE_OF_MINORITY = 16;
+    public static final long STANDARD_LOAN_PERIOD_WEEKS = 4L;
+    public static final int AGE_OF_MINORITY = 16;
 
     private Constant() {}
 }
